requirements.txt
joblib==1.1.0
machine==0.0.1
matrix_factorization==1.3
network==0.1
numpy==1.21.2
paho_mqtt==1.6.1
pandas==1.4.1
requests==2.27.1
scikit_learn==1.0.2
